Revert back to explicitly staging which services are public
By setting everything to public everything will get instantiated whether
it's used or not. This causes the list of things excluded to get very
long. Added benefit is we have a list of classes that get loaded through
a $container->get() method which is bad practice. These classes should
get autowired/injected in the relevant classes